feat(fuel-prices): refactor dashboard into modular components
Refactors the `FuelPriceDashboard` into smaller, reusable components to improve code structure, maintainability, and separation of concerns. The monolithic component has been broken down, with its logic and UI elements extracted into specialized components.

-   **`FuelPriceDashboard.tsx`**: Streamlined to manage fetching station data and controlling the visibility of the detail modal. It now renders a list of `FuelStationCard` components.

-   **`FuelStationCard.tsx` (New)**: A new component to display a summary for each fuel station. Each card shows current prices, price trends, and a mini 24-hour price history chart. Clicking a fuel type opens the detail modal.

-   **`FuelPriceChart.tsx` (New)**: Extracts all chart-related logic into a dedicated component. It handles fetching historical and forecast data, rendering various price lines (actual, trend, Holt-Winters), and provides interactive zoom and pan controls in "check mode".

-   **`FuelPriceDetailModal.tsx` (New)**: Encapsulates the entire detail view modal. This component contains the main `FuelPriceChart`, controls for selecting forecast types and time duration, toggles for line visibility, and a comprehensive help section.
